NetBeans Platform Project Started
I have started the RCP Project and created the `Northwind-Utilities` module, including the following:

* `com.pekinsoft.northwind.utils`
   - `Logger` class
   - `StringUtils` class
   - `.enums`
      * `SysExits` enumeration
   - `.exceptions`
      * `InvalidLoggingLevelException` class

I also modified the `Logger` class to no longer use the `Messaging` class that I had created. Instead, it now uses the `NotifyDescriptor` and `DialogDisplayer` for showing error messages from the logging methods.

Now, we just need to figure out how to get a handle on the `messages.log` that the NetBeans Platform logs to, so that we can have all logging to a single file. Otherwise, we will need to create a separate log for each class, which I don't want. However, that may be what we are stuck with.