Accounting Module Started
I have started working on mocking up the `Northwind-Accounting` module. I am going to create the visual elements to give us an idea of how the system will look.

I need to Open an Issue for the "Chart of Accounts" window, though. I have it set to be an explorer window, but it keeps opening in the editor area. I have tried changing it to a properties window, but that did not fix the problem either. I am not sure what to make of this. Maybe you could look at it @jkovalsky and see if I am missing something.
Signed-off-by:Sean Carrick <sean@pekinsoft.com>