GUI Moving Along
I have got the GUI in the Accounting Module moving along. I have now created the Chart of Accounts window as a Properties View, but it is not showing up in the Properties area for some reason, and the General Ledger window as an Editor View.

We need to figure out why the Chart of Accounts window will not appear in the Properties area, when that is what it is set up to be:

```java
@ConvertAsProperties(
        dtd = "-//com.pekinsoft.northwind.accounting.view//CoA//EN",
        autostore = false
)
@TopComponent.Description(
        preferredID = "CoATopComponent",
        iconBase = "com/pekinsoft/northwind/accounting/view/Clipboard.png",
        persistenceType = TopComponent.PERSISTENCE_ALWAYS
)
@TopComponent.Registration(mode = "properties", openAtStartup = true)
@ActionID(category = "Window", id = "com.pekinsoft.northwind.accounting.view.CoATopComponent")
@ActionReference(path = "Menu/Window" /*, position = 333 */)
@TopComponent.OpenActionRegistration(
        displayName = "#CTL_CoAAction",
        preferredID = "CoATopComponent"
)
@Messages({
    "CTL_CoAAction=CoA",
    "CTL_CoATopComponent=Chart of Accounts",
    "HINT_CoATopComponent=This is the Chart of Accounts"
})
```

As is clearly seen in the line `@TopComponent.Registration(mode = "properties", openAtStartup = true)`, this TopComponent is registered in the properties mode. However, it continues to display in the editor mode. This has me confused.

As well as the windows, I have created actions to display them. I placed the action for displaying the Chart of Accounts window under the menu 'Window/Optional Tools' and the action for the General Ledger under the 'View' menu. Also, I have given the General Ledger a toolbar button, located on the Performance toolbar node.

Other than this issue, the GUI is coming along nicely.
Signed-off-by:Sean Carrick <sean@pekinsoft.com>